added multi characters
its a litle finiky but i have added the option to purchase the other character and have made it to that the player gets coins over time :)

+++ GameEnginesAndLogicApp/Assets/Scripts/Kaz's Scripts/InGameCurrencySO.cs
@@ -8,5 +8,5 @@ public class InGameCurrencySO : ScriptableObject
 {
     public string currencyName;
     public Sprite currency;
-    public int currencyAmount;
+    public float currencyAmount;
 }
